Finalizing changes
